refactor: make VariantConverter private in settings.cc
