+++ libtransmission/variant.h
@@ -555,19 +555,4 @@ private:
     char const* end_ = nullptr;
 };
 
-namespace libtransmission
-{
-
-struct VariantConverter
-{
-public:
-    template<typename T>
-    static std::optional<T> load(tr_variant const& src);
-
-    template<typename T>
-    static tr_variant save(T const& val);
-};
-
-} // namespace libtransmission
-
 /* @} */
